src/servo/dom/bindings/codegen/test/test_sequence_wrapping.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=775852
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 775852</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=775852">Mozilla Bug 775852</a>
<p id="display"></p>
<div id="content" style="display: none">
  <canvas width="1" height="1" id="c"></canvas>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 775852 **/
function doTest() {
  try {
    var gl = $("c").getContext("experimental-webgl");
  } catch (e) {
    // No WebGL support on MacOS 10.5.  Just skip this test
    todo(false, "WebGL not supported");
    return;
  }
  var setterCalled = false;

  extLength = gl.getSupportedExtensions().length;
  ok(extLength > 0,
     "This test won't work right if we have no supported extensions");

  Object.defineProperty(Array.prototype, "0",
    {
      set: function(val) {
             setterCalled = true;
           }
    });

  // Test that our property got defined correctly
  var arr = []
  arr[0] = 5;
  is(setterCalled, true, "Setter should be called when setting prop on array");

  setterCalled = false;

  is(gl.getSupportedExtensions().length, extLength,
     "We should still have the same number of extensions");

  is(setterCalled, false,
     "Setter should not be called when getting supported extensions");
}
doTest();
</script>
</pre>
</body>
</html>
